Fix crash in CHbenCHmark queries
The logical planner stops on an assert, or crashes on a nullptr dereference for complicated joins. This is cause by a mis-assumption in the LocalJoin logic that did not assume a cartesian/repartition join before a reference join. This would cause currentJoinNode to be of type REFERENCE but without an anchorTable, instead of checking the join right before our current candidate we just skip if there is no anchorTable on the currentJoinNode.